Refactor asset providers (#383)
* Refactor asset provider to handle JS asset setup

* Move progress listener for flicker bug out of asset providers

* Pre-fetch assets to publish asset state on event broker
