+++ plugin/src/software/aws/toolkits/eclipse/amazonq/providers/assets/ToolkitLoginWebViewAssetProvider.java
@@ -11,15 +11,53 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Optional;
 
+import org.eclipse.swt.browser.Browser;
+import org.eclipse.swt.browser.BrowserFunction;
+
 import software.aws.toolkits.eclipse.amazonq.plugin.Activator;
+import software.aws.toolkits.eclipse.amazonq.telemetry.UiTelemetryProvider;
 import software.aws.toolkits.eclipse.amazonq.util.PluginUtils;
 import software.aws.toolkits.eclipse.amazonq.util.ThemeDetector;
 import software.aws.toolkits.eclipse.amazonq.util.WebviewAssetServer;
+import software.aws.toolkits.eclipse.amazonq.views.LoginViewActionHandler;
+import software.aws.toolkits.eclipse.amazonq.views.LoginViewCommandParser;
+import software.aws.toolkits.eclipse.amazonq.views.ViewActionHandler;
+import software.aws.toolkits.eclipse.amazonq.views.ViewCommandParser;
+import software.aws.toolkits.eclipse.amazonq.views.ViewConstants;
 
 public final class ToolkitLoginWebViewAssetProvider extends WebViewAssetProvider {
 
     private WebviewAssetServer webviewAssetServer;
     private static final ThemeDetector THEME_DETECTOR = new ThemeDetector();
+    private final ViewCommandParser commandParser;
+    private final ViewActionHandler actionHandler;
+
+    public ToolkitLoginWebViewAssetProvider() {
+        this.commandParser = new LoginViewCommandParser();
+        this.actionHandler = new LoginViewActionHandler();
+    }
+
+    @Override
+    public void injectAssets(final Browser browser) {
+        new BrowserFunction(browser, ViewConstants.COMMAND_FUNCTION_NAME) {
+            @Override
+            public Object function(final Object[] arguments) {
+                commandParser.parseCommand(arguments)
+                        .ifPresent(command -> actionHandler.handleCommand(command, browser));
+                return null;
+            }
+        };
+        new BrowserFunction(browser, "telemetryEvent") {
+            @Override
+            public Object function(final Object[] arguments) {
+                String clickEvent = (String) arguments[0];
+                UiTelemetryProvider.emitClickEventMetric("auth_" + clickEvent);
+                return null;
+            }
+        };
+
+        browser.setText(getContent().get());
+    }
 
     @Override
     public Optional<String> getContent() {
