+++ plugin/src/software/aws/toolkits/eclipse/amazonq/providers/assets/WebViewAssetProvider.java
@@ -5,7 +5,11 @@ package software.aws.toolkits.eclipse.amazonq.providers.assets;
 
 import java.util.Optional;
 
+import org.eclipse.swt.browser.Browser;
+
 public abstract class WebViewAssetProvider {
+
+    public abstract void injectAssets(final Browser browser);
 
     public abstract Optional<String> getContent();
 
@@ -29,6 +33,10 @@ public abstract class WebViewAssetProvider {
                     });
                 }
                 """;
+    }
+
+    protected final void disableBrowserContextMenu(final Browser browser) {
+        browser.execute("document.oncontextmenu = e => e.preventDefault();");
     }
 
 }
